Updated visualization functions, reorganizing code structure to simplify, extended 5000 entry defect log dataset
