+++ queueless-MC/BuildHistories.py
@@ -1,0 +1,113 @@
+import polars as pl
+import numpy as np
+import matplotlib.pyplot as plt
+from datetime import date
+from ..utils import VisualizationFunctions
+
+
+class BuildHistories:
+    def __init__(self, control_type, sub_log_df, sub_deltas_df, incoming_dict, outgoing_dict, empirical_dict):
+        self.control_type = control_type
+        self.sub_log_df_inc = sub_log_df.filter(pl.col('State') == "new")
+        self.sub_log_df_out = sub_log_df.filter(pl.col('State') == "closed")
+        self.sub_deltas_df = sub_deltas_df
+        self.incoming_dict = incoming_dict
+        self.outgoing_dict = outgoing_dict
+        self.empirical_dict = empirical_dict
+        self.VisualizationFunctions = VisualizationFunctions()
+
+    def update_figures(self):
+        """
+        Updates timeline figures tracking evolution of time deltas (per defect type)
+        """
+        fig, axs = plt.subplots(1, 4, figsize=(16, 4))
+        deltas = ['Delta_New_Assign', 'Delta_Assign_InProgress', 'Delta_InProgress_Closed', 'Delta_New_Closed']
+        ylabels = ['time (hrs)', None, None, None]
+        for index in range(4):
+            xdata = range(1,len(self.sub_deltas_df.select(deltas[index])) + 1)
+            ydata = self.sub_deltas_df.select(deltas[index])
+            fig, axs = self.VisualizationFunctions.visualize_timeline(fig,
+                                                                      axs,
+                                                                      index,
+                                                                      xdata,
+                                                                      ydata,
+                                                                      ylabel=ylabels[index],
+                                                                      title=deltas[index])
+    
+        fig.text(0.5, 0.95, f'{self.control_type}', ha='center', fontsize=12)
+        plt.savefig(f'figures/{self.control_type}_fig_{date.today()}.png', bbox_inches='tight')
+        plt.close()
+
+    def update_delta_distributions(self):
+        """
+        Updates empirical log data delta histograms (per defect type)
+        """
+        fig, axs = plt.subplots (1, 4, figsize=(16,4))
+        deltas = ['Delta_New_Assign', 'Delta_Assign_InProgress', 'Delta_InProgress_Closed', 'Delta_New_Closed']
+        for index in range(4):
+            data = self.empirical_dict[self.control_type][deltas[index].lower()]
+            fig, axs = self.VisualizationFunctions.visualize_histograms_theory(fig,
+                                                                               axs,
+                                                                               index,
+                                                                               data,
+                                                                               label=None,
+                                                                               title=deltas[index],
+                                                                               color='#C21445')
+        plt.savefig(f'figures/{self.control_type}_delta_histograms_{date.today()}.png', bbox_inches='tight')
+        plt.close()
+
+    def update_incoming(self):
+        """
+        Measures and stores incoming defects per hour (per defect type) 
+        
+        Returns
+            :incoming_dict (dict): tracks incoming defects per hour (per defect type)
+            :empirical_dict (dict): tracks incoming/outgoing and delta histograms from empirical data (per defect type)
+        """
+        for day in self.sub_log_df_inc["Date"].unique():
+            sub_log_df_inc_date = self.sub_log_df_inc.filter(pl.col('Date') == day)['Hour'].value_counts()
+            incoming_defects = [sub_log_df_inc_date.filter(pl.col('Hour') == i)['count'][0] if i in list(sub_log_df_inc_date.select('Hour'))[0] else 0 for i in range(24)]
+            self.incoming_dict[self.control_type].update({day: incoming_defects})
+            self.empirical_dict[self.control_type]['incoming_per_hour'] = self.empirical_dict[self.control_type]['incoming_per_hour'] + self.incoming_dict[self.control_type][day]
+        return self.incoming_dict, self.empirical_dict
+    
+    def update_outgoing(self):
+        """
+        Measures and stores outgoing defects per hour (per defect type)
+        
+        Returns
+            :outgoing_dict (dict): tracks outgoing defects per hour (per defect type)
+            :empirical_dict (dict): tracks incoming/outgoing and delta histograms from empirical data (per defect type)
+        """
+        for day in self.sub_log_df_out["Date"].unique(): # will never be empty 
+            sub_log_df_out_date = self.sub_log_df_out.filter(pl.col('Date') == day)['Hour'].value_counts()
+            outgoing_defects = [sub_log_df_out_date.filter(pl.col('Hour') == i)['count'][0] if i in list(sub_log_df_out_date.select('Hour'))[0] else 0 for i in range(24)]
+            self.outgoing_dict[self.control_type].update({day: outgoing_defects})
+            self.empirical_dict[self.control_type]['outgoing_per_hour'] = self.empirical_dict[self.control_type]['outgoing_per_hour'] + self.outgoing_dict[self.control_type][day]
+        return self.outgoing_dict, self.empirical_dict
+    
+    def update_incoming_outgoing_distributions(self):
+        """
+        Updates empirical log data incoming + outgoing histograms (per defect type)
+        """
+        fig, axs = plt.subplots (1, 2, figsize=(16,4))
+        # incoming_per_hour histogram
+        data_incoming = np.array(self.empirical_dict[self.control_type]['incoming_per_hour'])
+        fig, axs = self.VisualizationFunctions.visualize_histograms_theory(fig,
+                                                                           axs,
+                                                                           0,
+                                                                           data_incoming,
+                                                                           label='incoming',
+                                                                           title=f'{self.control_type} incoming [/hr]',
+                                                                           color='#C21445')
+        # outgoing_per_hour histogram
+        data_outgoing = np.array(self.empirical_dict[self.control_type]['outgoing_per_hour'])
+        fig, axs = self.VisualizationFunctions.visualize_histograms_theory(fig,
+                                                                           axs,
+                                                                           1,
+                                                                           data_outgoing,
+                                                                           label='outgoing',
+                                                                           title=f'{self.control_type} outgoing [/hr]',
+                                                                           color='#C21445')
+        plt.savefig(f'figures/{self.control_type}_histograms_{date.today()}.png', bbox_inches='tight')
+        plt.close()
