+++ queueless-MC/ControlTypes.py
@@ -1,4 +1,4 @@
-from DeltaSimulation import DefectType
+import DefectType
 
 class ACC03(DefectType):
     def __init__(self, sub_log_df_ACC03):
